Fix type mismatches in C++ standard codec (flutter/engine#9112)
There were some implicit casts in the standard codec implementation that
didn't show up on Linux, but do on Windows.

+++ engine/src/flutter/shell/platform/common/cpp/client_wrapper/standard_codec_serializer.h
@@ -30,10 +30,10 @@ class StandardCodecSerializer {
 
  protected:
   // Reads the variable-length size from the current position in |stream|.
-  uint32_t ReadSize(ByteBufferStreamReader* stream) const;
+  size_t ReadSize(ByteBufferStreamReader* stream) const;
 
   // Writes the variable-length size encoding to |stream|.
-  void WriteSize(uint32_t size, ByteBufferStreamWriter* stream) const;
+  void WriteSize(size_t size, ByteBufferStreamWriter* stream) const;
 
   // Reads a fixed-type list whose values are of type T from the current
   // position in |stream|, and returns it as the corresponding EncodableValue.
